[PAINT-162] Moved ViewActions from espressoutils to UiInteractions; moved matchers from espressoutils to UiMatcher utils class; added position coordinates provider class
